requirements.txt
fastapi
python-dotenv
uvicorn
fastapi-cache2
git+https://github.com/Rapptz/asqlite
